refactor: extract useActivityChat hook from ActivityTimeline
채팅 패널 orchestration 로직(useChatPanel, useChatPanelSync 조율, handleOpenChat)을 useActivityChat 훅으로 분리했습니다. ValidationState, ValidationTarget 타입도 훅에서 export해 중복을 제거했습니다.
